+++ lib/cassandra/driver.rb
@@ -142,7 +142,7 @@ module Cassandra
     end
 
     let(:port)                      { 9042 }
-    let(:protocol_version)          { 4 }
+    let(:protocol_version)          { nil }
     let(:connect_timeout)           { 10 }
     let(:ssl)                       { false }
     let(:logger)                    { NullLogger.new }